Split the existing pipeline into some combined stages
