+++ app/src/main/java/dk/scuffed/whiteboardapp/pipeline/StageCombinations/FullCornerDetection.kt
@@ -1,0 +1,58 @@
+package dk.scuffed.whiteboardapp.pipeline.StageCombinations
+
+import android.content.Context
+import android.graphics.Bitmap
+import dk.scuffed.whiteboardapp.pipeline.Pipeline
+import dk.scuffed.whiteboardapp.pipeline.stages.GLOutputStage
+import dk.scuffed.whiteboardapp.pipeline.stages.PointsOutputStage
+import dk.scuffed.whiteboardapp.pipeline.stages.bitmap_process_stages.FramebufferToBitmapStage
+import dk.scuffed.whiteboardapp.pipeline.stages.bitmap_process_stages.OpenCVLineDetectionStage
+import dk.scuffed.whiteboardapp.pipeline.stages.lines_stages.BiggestSquareStage
+import dk.scuffed.whiteboardapp.pipeline.stages.lines_stages.LinesAngleDiscriminatorStage
+
+/**
+ * Does corner detection using hough line transforms and finding the biggest quadrilateral.
+ * Preprocesses the image using canny edge detection.
+ * @return points found
+ */
+internal fun fullCornerDetection(context: Context, inputStage: GLOutputStage, pipeline: Pipeline): PointsOutputStage {
+    val edges = fullCannyEdgeDetection(
+        context,
+        inputStage,
+        pipeline
+    )
+
+    val edgesBitmapStage = FramebufferToBitmapStage(
+        edges.frameBufferInfo,
+        Bitmap.Config.ARGB_8888,
+        pipeline
+    )
+
+    val openCVLineDetectionStage = OpenCVLineDetectionStage(
+        edgesBitmapStage,
+        150,
+        pipeline
+    )
+
+    val verticalLinesAngleDiscriminatorStage = LinesAngleDiscriminatorStage(
+        openCVLineDetectionStage,
+        -(Math.PI / 4.0f).toFloat(),
+        (Math.PI / 4.0f).toFloat(),
+        pipeline
+    )
+
+    val horizontalLinesAngleDiscriminatorStage = LinesAngleDiscriminatorStage(
+        openCVLineDetectionStage,
+        (Math.PI / 4.0f).toFloat(),
+        (Math.PI / 2.0f + Math.PI / 4.0f).toFloat(),
+        pipeline
+    )
+
+    val biggestSquareStage = BiggestSquareStage(
+        horizontalLinesAngleDiscriminatorStage,
+        verticalLinesAngleDiscriminatorStage,
+        pipeline
+    )
+
+    return biggestSquareStage
+}
